Move all type definitions to a new file types.go
